app/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "InstantReserve"
version = "0.0.1"
description = 'A website that consolidates different business reservations into one place'
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "Rosa Miranda", email = "rmirandagomez4149@sdsu.edu" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "fastapi[all]>=0.115.4",
  "passlib[bcrypt]>=1.7.4",
  "pydantic-settings>=2.6.0",
  "pyjwt>=2.9.0",    
  "pylance>=0.19.2",
  "python-dotenv>=1.0.1",
  "sqlalchemy>=2.0.36",
  "sqlmodel>=0.0.22",
]

[project.urls]
Documentation = "https://github.com/U.N. Owen/app#readme"
Issues = "https://github.com/U.N. Owen/app/issues"
Source = "https://github.com/U.N. Owen/app"

[tool.hatch.version]
path = "src/app/__about__.py"

[tool.hatch.envs.default]
python = "3.12"
installer = "uv"
path = ".venv"

[tool.hatch.envs.default.scripts]
dev = "uvicorn app.main:app --reload"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.envs.default.env-vars]
PYTHONASYNCIODEBUG = "1"

[tool.hatch.envs.types]
installer = "uv"
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/app tests}"

[tool.coverage.run]
source_pkgs = ["app", "tests"]
branch = true
parallel = true
omit = [
  "src/app/__about__.py",
]

[tool.coverage.paths]
app = ["src/app", "*/app/src/app"]
tests = ["tests", "*/app/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]